telegrambot/requirements.txt
aiogram>=3.0.0
httpx>=0.23.0
aiohttp>=3.8.0
python-dotenv>=1.0.0